Validation pattern created.
(Let javascript syntax bugs go to the mothers p*y -- translate to Croatian for full experience)
At this point only age validation for a part of the form in EditInfo is working.

EditInfo consists of 3 files:
	- UI Component (find it inside client private components folder).
	- Data load container (find it inside client private containers folder).
	- Actions it can take (find it inside client actions).

Data load container exists to create a UI Component and load it with data. Data consists of:
	- Error object from LocalState regarding that UI Component.
	- Actions (not all actions, only certain functions).

After the UI Component is created, the user will interact with it. For this example let's say
the user enters some letters into the "age" part of the form. After the user removes focus from
the "age", the UI Component will fire the "onBlurHandler" and call the validation function that
it was injected with (in the Data load container). The "ageValidation" function will notice that
the given value is not a number and it will set the "LocalState" with an error object. "ageValidation"
function will supply the error object with a "status" and a "message" regarding what happened.

	NOTE: In case of no errors you must set LocalState with an error of success and an empty message.
	NOTE NOTE: When creating an action file for a UI Component, you MUST create a "clearError" function
	and inject it into the UI Component. Also you must return that same function inside the "composer" part
	of the "data load container"! If you forget this, error object inside LocalState won't be deleted (ever!).
	Example of a clearError function can be found in the action file for EditInfo.

Since the "LocalState" is a reactive source and the data load container is a reactive computation, the
data load container will run again and supply the UI Component with a new error object. The UI component
will render differently depending on the error object.