Refactor containers CustomizeCharacter and ProfileSetup to use custom Loading component which is an empty div tag. Also call "NProgress.done()" inside the container when subscription/s is/are ready.
This solves a bug that appeared in the undo mechanism inside CustomizeCharacter after commit 9afd405.

Using custom Loading inside container instead of rendering empty div inside component assures that all subscriptions will be ready when the constructor for the component is called. Component constructor is called only once. Everything after that until component unmounts is just container compose and component render.

Undo mechanism relied on the fact that the component constructor will be called only once but it also relied on the fact that the subscriptions will be ready at the time of the construction. Without a custom loading component inside it's container the subscriptions are never going to be ready inside components constructor.

Since react-komposer offers a clean solution I decided to utilize it.